improved test

+++ tests/base_communication_test.py
@@ -7,6 +7,19 @@ import threading
 file_name_to_slave = "signal_to_slave.dat"
 file_name_from_slave = "signal_from_slave.dat"
 
+# map signal - [size, case-Id]
+# NOTE needs to be consistent with C++
+signal_map = {
+    "data_0" : 25,
+    "data_1" : 25,
+    "data_2" : 25,
+    "data_3" : 25,
+    "data_4" : 1E10,
+    "data_5" : 1E10,
+    "data_6" : 1E10,
+    "data_7" : 1E10,
+}
+
 class WrapperClass(object):
     # wrapping in an extra class to avoid discovery of the base-test
     # see https://stackoverflow.com/a/25695512
@@ -34,39 +47,50 @@ class WrapperClass(object):
             pass
 
         def test_send_receive_int_once_small(self):
-            pass
+            self.__ExecuteSendTest("data_0")
 
         def test_send_receive_double_once_small(self):
-            pass
+            self.__ExecuteSendTest("data_1")
 
         def test_send_receive_int_once_large(self):
-            pass
+            self.__ExecuteSendTest("data_2")
 
         def test_send_receive_double_once_large(self):
-            pass
+            self.__ExecuteSendTest("data_3")
 
         def test_send_receive_int_multiple_times_small(self):
-            pass
+            self.__ExecuteSendTestMultiple("data_4")
 
         def test_send_receive_double_multiple_times_small(self):
-            pass
+            self.__ExecuteSendTestMultiple("data_5")
 
         def test_send_receive_int_multiple_times_large(self):
-            pass
+            self.__ExecuteSendTestMultiple("data_6")
 
         def test_send_receive_double_multiple_times_large(self):
-            pass
+            self.__ExecuteSendTestMultiple("data_7")
+
+        def __ExecuteSendTestMultiple(self, data_signal):
+            for _ in range(100):
+                self.__ExecuteSendTest(data_signal)
+
+        def __ExecuteSendTest(self, data_signal):
+            self.__CommunicateSignalToSlaveProcess(self.connection_name, self.comm_name, data_signal)
+            data_size = signal_map[data_signal]
+            self.master_comm.Receive(int(data_size), int(data_signal[5:]))
+
 
         @classmethod
         def __CommunicateSignalToSlaveProcess(self, connection_name, comm_name, signal):
             WaitForFileToBeRemoved(file_name_to_slave)
 
-            with open(file_name_to_slave, 'w') as signal_file:
+            with open("."+file_name_to_slave, 'w') as signal_file:
                 signal_file.write(connection_name)
                 signal_file.write("\n")
                 signal_file.write(comm_name)
                 signal_file.write("\n")
                 signal_file.write(signal)
+            os.rename("."+file_name_to_slave, file_name_to_slave) # making "visible"
 
         def tearDown(self):
             self.__CommunicateSignalToSlaveProcess(self.connection_name, self.comm_name, "disconnect")
@@ -114,8 +138,13 @@ class BaseCommunicationTestDataSender(object):
                 self.slave_comm = None
                 break
 
+            elif signal in signal_map:
+                self.__CheckConnection(connection_name, comm_name)
+                data_size = signal_map[signal]
+                self.slave_comm.Send(int(data_size), int(signal[5:]))
+
             else:
-                raise Exception('Signal "{}" not recognized!')
+                raise Exception('Signal "{}" not recognized!'.format(signal))
 
     def ExecuteInThread(self):
         self.my_thread = threading.Thread(target=self.Execute)
@@ -155,6 +184,9 @@ def ReceiveSignalFromMasterProcess():
 
     with open(file_name_to_slave, 'r') as signal_file:
         read_signal_info = signal_file.read().splitlines()
+
+    if not len(read_signal_info) == 3:
+        raise Exception("Wrong signal received:", read_signal_info)
 
     os.remove(file_name_to_slave)
